float => double

+++ src/chart_common.h
@@ -141,7 +141,7 @@ namespace Chart {
   );
 
   void MakeColorVisible(
-    SVG::Color* color, SVG::Color* bg_color, float min_visibility = 0.3
+    SVG::Color* color, SVG::Color* bg_color, double min_visibility = 0.3
   );
 
   // Detect if the given string most likely contains only normal width UTF-8
